+++ src/inference/inference_mxnet_async_mode.py
@@ -1,14 +1,11 @@
 import argparse
 import json
 import logging as log
-import os
 import sys
 import traceback
-import warnings
 from pathlib import Path
 from time import time
 
-import gluoncv
 import mxnet
 
 import postprocessing_data as pp
@@ -16,7 +13,7 @@ from io_adapter import IOAdapter
 from io_model_wrapper import MXNetIOModelWrapper
 from reporter.report_writer import ReportWriter
 from transformer import MXNetTransformer
-from mxnet_auxiliary import prepare_output
+from mxnet_auxiliary import *
 
 
 def cli_argument_parser():
@@ -158,71 +155,6 @@ def cli_argument_parser():
     args = parser.parse_args()
 
     return args
-
-
-def get_device_to_infer(device):
-    log.info('Get device for inference')
-    if device == 'CPU':
-        log.info(f'Inference will be executed on {device}')
-        return mxnet.cpu()
-    elif device == 'NVIDIA_GPU':
-        log.info(f'Inference will be executed on {device}')
-        return mxnet.gpu()
-    else:
-        log.info(f'The device {device} is not supported')
-        raise ValueError('The device is not supported')
-
-
-def load_network_gluon(model_json, model_params, context, input_name):
-    log.info(f'Deserializing network from file ({model_json}, {model_params})')
-    with warnings.catch_warnings():
-        warnings.simplefilter('ignore')
-        deserialized_net = mxnet.gluon.nn.SymbolBlock.imports(
-            model_json, [input_name], model_params, ctx=context)
-    return deserialized_net
-
-
-def load_network_gluon_model_zoo(model_name, hybrid, context, save_model, path_save_model):
-    log.info(f'Loading network \"{model_name}\" from GluonCV model zoo')
-    net = gluoncv.model_zoo.get_model(model_name, pretrained=True, ctx=context)
-
-    if save_model is True:
-        log.info(f'Saving model \"{model_name}\" to \"{path_save_model}\"')
-        if path_save_model is None:
-            path_save_model = os.getcwd()
-        path_save_model = os.path.join(path_save_model, model_name)
-        if not os.path.exists(path_save_model):
-            os.mkdir(path_save_model)
-        gluoncv.utils.export_block(os.path.join(path_save_model, model_name), net,
-                                   preprocess=None, layout='CHW', ctx=context)
-
-    log.info(f'Info about the network:\n{net}')
-
-    log.info(f'Hybridizing model to accelerate inference: {hybrid}')
-    if hybrid is True:
-        net.hybridize()
-    return net
-
-
-def create_dict_for_transformer(args):
-    dictionary = {
-        'channel_swap': args.channel_swap,
-        'mean': args.mean,
-        'std': args.std,
-        'norm': args.norm,
-        'input_shape': args.input_shape,
-        'batch_size': args.batch_size,
-    }
-    return dictionary
-
-
-def create_dict_for_modelwrapper(args):
-    dictionary = {
-        'input_name': args.input_name,
-        'input_shape': [args.batch_size] + args.input_shape[1:4],
-        'model_name': args.model_name,
-    }
-    return dictionary
 
 
 def inference_mxnet(net, num_iterations, get_slice, input_name):
